refactor(inventory): fix phase 1 module boundaries for page fetch
