+++ src/features/liferay/inventory/liferay-inventory-page-fetch-http.ts
@@ -1,0 +1,15 @@
+import type {LiferayGateway} from '../liferay-gateway.js';
+
+export async function safeGatewayGet<T>(
+  gateway: LiferayGateway,
+  requestPath: string,
+  label: string,
+  requestOptions?: {headers?: Record<string, string>},
+): Promise<{ok: boolean; data: T | null}> {
+  try {
+    const data = await gateway.getJson<T>(requestPath, label, requestOptions);
+    return {ok: true, data};
+  } catch {
+    return {ok: false, data: null};
+  }
+}
