+++ src/features/liferay/inventory/liferay-inventory-page-fetch.ts
@@ -1,10 +1,7 @@
 /* eslint-disable max-lines -- inventory orchestration intentionally consolidated during active refactor */
 import type {AppConfig} from '../../../core/config/load-config.js';
-import path from 'node:path';
-import fs from 'fs-extra';
 import type {LiferayApiClient} from '../../../core/http/client.js';
 import {firstNonEmptyString, firstPositiveNumber, toBoolean, toBooleanOrFalse} from '../../../core/utils/coerce.js';
-import {trimLeadingSlash} from '../../../core/utils/text.js';
 import {LiferayErrors} from '../errors/index.js';
 import type {ResolvedSite} from './liferay-inventory-shared.js';
 import {
@@ -17,18 +14,10 @@ import type {LiferayGateway} from '../liferay-gateway.js';
 import {buildJournalArticleAdminUrls, buildLayoutAdminUrls} from '../page-layout/liferay-page-admin-urls.js';
 import {
   asRecord,
-  assignOptionalFiniteNumber,
-  assignOptionalString,
-  assignOptionalNumber,
-  assignOptionalBoolean,
   collectPageElements,
-  hasJsonWsException,
-  summarizeContentFields,
-  firstString,
   type ContentStructureSummary,
   type JournalArticleSummary,
   type PageFragmentEntry,
-  type StructuredContent,
 } from './liferay-inventory-page-assemble.js';
 import {KNOWN_LOCALES} from './liferay-inventory-page-url.js';
 import type {
@@ -36,18 +25,18 @@ import type {
   PagePortletSummary,
   ResolvedRegularLayoutPage,
 } from './liferay-inventory-page.js';
-import {resolveSiteToken, resolveFragmentsBaseDir, resolveArtifactSiteDir} from '../resource/liferay-resource-paths.js';
-import {
-  buildResourceSiteChain,
-  fetchGroupInfo,
-  listDdmTemplates,
-  resolveResourceSite,
-} from '../resource/liferay-resource-shared.js';
-import {matchesDdmTemplate} from '../liferay-identifiers.js';
 import {classNameIdLookupCache} from '../lookup-cache.js';
+import {resolveDisplayPageArticle, resolveStructuredContentData} from './liferay-inventory-page-fetch-article.js';
+import {safeGatewayGet} from './liferay-inventory-page-fetch-http.js';
+import {fetchComponentPageData} from './liferay-inventory-page-fetch-components.js';
+import {enrichFragmentEntryExportPaths} from './liferay-inventory-page-fetch-fragments.js';
+import {
+  buildJournalArticleSummary,
+  collectLayoutContentStructures,
+  collectLayoutJournalArticles,
+} from './liferay-inventory-page-fetch-journal.js';
 
 type LayoutMatch = {layout: Layout; locale: string | null};
-type ArticleRef = {articleId: string; groupId: number; ddmTemplateKey?: string};
 
 const CLASS_NAME_LAYOUT = 'com.liferay.portal.kernel.model.Layout';
 const CLASS_NAME_JOURNAL_ARTICLE = 'com.liferay.journal.model.JournalArticle';
@@ -73,86 +62,6 @@ export async function fetchSiteRootInventory(
       type: layout.type ?? '',
     })),
   };
-}
-
-/**
- * Resolve display page article with fallback chain:
- * 1. Try headless-delivery structured content by friendlyUrlPath
- * 2. Fall back to JSONWS get-article-by-url-title
- * Returns: { article, jsonwsArticle, articleRef }
- */
-async function resolveDisplayPageArticle(
-  gateway: LiferayGateway,
-  siteId: number,
-  urlTitle: string,
-): Promise<{article: StructuredContent; jsonwsArticle: Record<string, unknown> | null; articleRef: ArticleRef}> {
-  const filter = encodeURIComponent(`friendlyUrlPath eq '${urlTitle}'`);
-  const response = await safeGatewayGet<{items?: StructuredContent[]}>(
-    gateway,
-    `/o/headless-delivery/v1.0/sites/${siteId}/structured-contents?filter=${filter}&pageSize=1`,
-    'site-structured-contents',
-  );
-  let article: StructuredContent | undefined = response.ok ? response.data?.items?.[0] : undefined;
-  let jsonwsArticle = await fetchJournalArticleByUrlTitle(gateway, siteId, urlTitle);
-
-  // Fallback: try JSONWS when headless delivery returns nothing
-  if (!article && jsonwsArticle) {
-    article = {
-      id: Number(jsonwsArticle.resourcePrimKey ?? jsonwsArticle.id ?? -1) || undefined,
-      key: String(jsonwsArticle.articleId ?? ''),
-      title: String(jsonwsArticle.titleCurrentValue ?? jsonwsArticle.title ?? ''),
-      friendlyUrlPath: urlTitle,
-      contentStructureId: Number(jsonwsArticle.contentStructureId ?? -1) || undefined,
-    };
-  }
-
-  if (!article) {
-    throw LiferayErrors.inventoryError(
-      `No structured content found with friendlyUrlPath=${urlTitle}. Verify the article URL title and site visibility, or confirm JSONWS/headless permissions for this OAuth client.`,
-    );
-  }
-
-  const articleRef: ArticleRef = {
-    articleId: String(article.key ?? jsonwsArticle?.articleId ?? ''),
-    groupId: siteId,
-    ...(firstString(jsonwsArticle?.ddmTemplateKey) ? {ddmTemplateKey: firstString(jsonwsArticle?.ddmTemplateKey)} : {}),
-  };
-
-  // Fetch JSONWS article if we don't have it yet
-  if (!jsonwsArticle && articleRef.articleId) {
-    jsonwsArticle = await fetchLatestJournalArticle(gateway, siteId, articleRef.articleId);
-  }
-
-  return {article, jsonwsArticle, articleRef};
-}
-
-/**
- * Resolve structured content data with fallback chain:
- * 1. Try by UUID (preferred, more complete)
- * 2. Fall back to by ID
- */
-async function resolveStructuredContentData(
-  gateway: LiferayGateway,
-  siteId: number,
-  article: StructuredContent,
-  jsonwsArticle: Record<string, unknown> | null,
-): Promise<StructuredContent | null> {
-  let structuredContent: StructuredContent | null = null;
-  const uuid = firstString(jsonwsArticle?.uuid);
-
-  if (uuid) {
-    structuredContent = await fetchStructuredContentByUuid(gateway, siteId, uuid);
-  }
-
-  // Fallback to fetch by ID if UUID didn't work
-  if (!structuredContent) {
-    const structuredContentId = article.id && article.id > 0 ? article.id : -1;
-    if (structuredContentId > 0) {
-      structuredContent = await fetchStructuredContentById(gateway, structuredContentId);
-    }
-  }
-
-  return structuredContent;
 }
 
 export async function fetchDisplayPageInventory(
@@ -893,706 +802,4 @@ async function findLayoutByPlidRecursive(
   }
 
   return null;
-}
-
-async function fetchSitePageElement(
-  gateway: LiferayGateway,
-  siteId: number,
-  friendlyUrl: string,
-): Promise<Record<string, unknown> | null> {
-  const slug = trimLeadingSlash(friendlyUrl);
-  const response = await safeGatewayGet<Record<string, unknown>>(
-    gateway,
-    `/o/headless-delivery/v1.0/sites/${siteId}/site-pages/${encodeURIComponent(slug)}?fields=pageDefinition`,
-    'fetch-site-page-element',
-  );
-  if (!response.ok) {
-    return null;
-  }
-  return asRecord(asRecord(response.data).pageDefinition).pageElement as Record<string, unknown> | null;
-}
-
-async function tryFetchSitePageMetadata(
-  gateway: LiferayGateway,
-  siteId: number,
-  friendlyUrl: string,
-): Promise<Record<string, unknown> | null> {
-  try {
-    const slug = trimLeadingSlash(friendlyUrl);
-    const response = await safeGatewayGet<Record<string, unknown>>(
-      gateway,
-      `/o/headless-delivery/v1.0/sites/${siteId}/site-pages/${encodeURIComponent(slug)}?nestedFields=taxonomyCategoryBriefs`,
-      'fetch-site-page-metadata',
-    );
-    if (!response.ok || !response.data) {
-      return null;
-    }
-    return asRecord(response.data);
-  } catch {
-    return null;
-  }
-}
-
-async function tryFetchFragmentEntryLinks(
-  gateway: LiferayGateway,
-  groupId: number,
-  plid: number,
-): Promise<Array<Record<string, unknown>>> {
-  if (plid <= 0) {
-    return [];
-  }
-  const response = await safeGatewayGet<Array<Record<string, unknown>>>(
-    gateway,
-    `/api/jsonws/fragment.fragmententrylink/get-fragment-entry-links?groupId=${groupId}&plid=${plid}`,
-    'fetch-fragment-entry-links',
-  );
-  return response.ok && Array.isArray(response.data) ? response.data : [];
-}
-
-async function collectLayoutJournalArticles(
-  gateway: LiferayGateway,
-  config: AppConfig,
-  apiClient: LiferayApiClient,
-  defaultGroupId: number,
-  fragmentEntryLinks: Array<Record<string, unknown>>,
-): Promise<JournalArticleSummary[]> {
-  const refs = extractArticleRefs(fragmentEntryLinks, defaultGroupId);
-  const result: JournalArticleSummary[] = [];
-
-  for (const ref of refs.values()) {
-    result.push(await buildJournalArticleSummary(gateway, config, apiClient, ref));
-  }
-
-  return result;
-}
-
-async function buildJournalArticleSummary(
-  gateway: LiferayGateway,
-  config: AppConfig,
-  apiClient: LiferayApiClient,
-  ref: ArticleRef,
-  options?: {
-    article?: Record<string, unknown> | null;
-    structuredContent?: StructuredContent | null;
-    fallbackSite?: ResolvedSite;
-    fallbackTitle?: string;
-    fallbackContentStructureId?: number;
-    includeHeadlessInventoryFields?: boolean;
-  },
-): Promise<JournalArticleSummary> {
-  const article = options?.article ?? (await fetchLatestJournalArticle(gateway, ref.groupId, ref.articleId));
-  const articleSite =
-    (await safeFetchGroupInfo(config, ref.groupId, {apiClient, gateway})) ??
-    (options?.fallbackSite
-      ? {
-          friendlyUrl: options.fallbackSite.friendlyUrlPath,
-          name: options.fallbackSite.name,
-          parentGroupId: -1,
-        }
-      : null);
-  const ddmTemplateKey = ref.ddmTemplateKey ?? firstString(article?.ddmTemplateKey);
-  const summary: JournalArticleSummary = {
-    groupId: ref.groupId,
-    ...(articleSite?.friendlyUrl ? {siteFriendlyUrl: articleSite.friendlyUrl} : {}),
-    ...(articleSite?.name ? {siteName: articleSite.name} : {}),
-    articleId: ref.articleId,
-    title: String(article?.titleCurrentValue ?? article?.title ?? options?.fallbackTitle ?? ref.articleId),
-    ddmStructureKey: String(article?.ddmStructureKey ?? ''),
-    ...(ddmTemplateKey ? {ddmTemplateKey} : {}),
-    ...(options?.fallbackContentStructureId ? {contentStructureId: Number(options.fallbackContentStructureId)} : {}),
-  };
-
-  // 1. Try to enrich with headless-delivery using UUID from JSONWS
-  let structuredContent = options?.structuredContent ?? null;
-  const uuid = firstString(article?.uuid);
-  if (!structuredContent && uuid) {
-    structuredContent = await fetchStructuredContentByUuid(gateway, ref.groupId, uuid);
-  }
-
-  // 2. If UUID lookup failed, try by ID as fallback
-  if (!structuredContent) {
-    const structuredContentId = Number(article?.id ?? article?.resourcePrimKey ?? -1);
-    if (structuredContentId > 0) {
-      structuredContent = await fetchStructuredContentById(gateway, structuredContentId);
-    }
-  }
-
-  // 3. Enrich summary with headless-delivery data (or JSONWS fallback if headless-delivery hasn't indexed yet)
-  if (structuredContent) {
-    if (options?.includeHeadlessInventoryFields) {
-      enrichJournalArticleWithStructuredContent(summary, structuredContent, ddmTemplateKey);
-    }
-    if (structuredContent.contentStructureId) {
-      summary.contentStructureId = Number(structuredContent.contentStructureId);
-    }
-    const contentFields = summarizeContentFields(structuredContent.contentFields);
-    if (contentFields.length > 0) {
-      summary.contentFields = contentFields;
-    }
-  } else if (article) {
-    // Fallback: Use JSONWS data when headless-delivery hasn't indexed yet
-    // Extract contentStructureId from JSONWS article
-    const jsonwsContentStructureId = Number(article.contentStructureId ?? -1);
-    if (jsonwsContentStructureId > 0) {
-      summary.contentStructureId = jsonwsContentStructureId;
-    }
-  }
-
-  // 4. Resolve structure key if not already available
-  if (!summary.ddmStructureKey && summary.contentStructureId) {
-    const contentStructure = await fetchContentStructureById(gateway, summary.contentStructureId);
-    const contentStructureKey = inferContentStructureKey(contentStructure);
-    if (contentStructureKey) {
-      summary.ddmStructureKey = contentStructureKey;
-    }
-  }
-
-  // 5. Resolve structure site and export path
-  if (articleSite?.friendlyUrl && summary.ddmStructureKey) {
-    const structureSite = await resolveStructureSiteByKey(
-      gateway,
-      config,
-      apiClient,
-      articleSite.friendlyUrl,
-      summary.ddmStructureKey,
-    );
-    if (structureSite) {
-      summary.ddmStructureSiteFriendlyUrl = structureSite.siteFriendlyUrl;
-      summary.structureExportPath = buildStructureExportPath(
-        config,
-        structureSite.siteFriendlyUrl,
-        summary.ddmStructureKey,
-      );
-    }
-  }
-
-  // 6. Resolve template site and export path
-  if (articleSite?.friendlyUrl && ddmTemplateKey) {
-    const templateSite = await resolveTemplateSiteByKey(config, articleSite.friendlyUrl, ddmTemplateKey, {
-      apiClient,
-      gateway,
-    });
-    if (templateSite) {
-      summary.ddmTemplateSiteFriendlyUrl = templateSite;
-      summary.templateExportPath = buildTemplateExportPath(config, templateSite, ddmTemplateKey);
-    }
-  }
-
-  return summary;
-}
-
-async function collectLayoutContentStructures(
-  gateway: LiferayGateway,
-  config: AppConfig,
-  apiClient: LiferayApiClient,
-  journalArticles: JournalArticleSummary[],
-): Promise<ContentStructureSummary[]> {
-  const seen = new Set<number>();
-  const result: ContentStructureSummary[] = [];
-
-  for (const article of journalArticles) {
-    const contentStructureId = Number(article.contentStructureId ?? -1);
-    if (contentStructureId <= 0 || seen.has(contentStructureId)) {
-      continue;
-    }
-    seen.add(contentStructureId);
-    const response = await safeGatewayGet<Record<string, unknown>>(
-      gateway,
-      `/o/headless-delivery/v1.0/content-structures/${contentStructureId}`,
-      'fetch-content-structure',
-    );
-    if (!response.ok) {
-      continue;
-    }
-    const key = inferContentStructureKey(response.data) || article.ddmStructureKey;
-    const siteFriendlyUrl = article.ddmStructureSiteFriendlyUrl ?? article.siteFriendlyUrl;
-    result.push({
-      contentStructureId,
-      ...(key ? {key} : {}),
-      name: String(response.data?.name ?? ''),
-      ...(siteFriendlyUrl ? {siteFriendlyUrl} : {}),
-      ...(siteFriendlyUrl && key ? {exportPath: buildStructureExportPath(config, siteFriendlyUrl, key)} : {}),
-    });
-  }
-
-  return result;
-}
-
-function extractArticleRefs(
-  fragmentEntryLinks: Array<Record<string, unknown>>,
-  defaultGroupId: number,
-): Map<string, ArticleRef> {
-  const refs = new Map<string, ArticleRef>();
-
-  for (const link of fragmentEntryLinks) {
-    const editableValues = String(link.editableValues ?? '').trim();
-    if (!editableValues || editableValues === '{}') {
-      continue;
-    }
-    try {
-      collectArticleRefsFromValue(JSON.parse(editableValues), refs, defaultGroupId);
-    } catch {
-      // Ignore invalid fragment editable values.
-    }
-  }
-
-  return refs;
-}
-
-function collectArticleRefsFromValue(value: unknown, refs: Map<string, ArticleRef>, defaultGroupId: number): void {
-  if (Array.isArray(value)) {
-    for (const item of value) {
-      collectArticleRefsFromValue(item, refs, defaultGroupId);
-    }
-    return;
-  }
-
-  const record = asRecord(value);
-  if (Object.keys(record).length === 0) {
-    return;
-  }
-
-  const directPrefsMap = asRecord(record.portletPreferencesMap);
-  const nestedPrefsMap = asRecord(asRecord(record.configuration).portletPreferencesMap);
-  collectArticleRefFromPreferences(directPrefsMap, refs, defaultGroupId);
-  collectArticleRefFromPreferences(nestedPrefsMap, refs, defaultGroupId);
-
-  for (const item of Object.values(record)) {
-    collectArticleRefsFromValue(item, refs, defaultGroupId);
-  }
-}
-
-function collectArticleRefFromPreferences(
-  prefsMap: Record<string, unknown>,
-  refs: Map<string, ArticleRef>,
-  defaultGroupId: number,
-): void {
-  const articleId = firstString(prefsMap.articleId);
-  if (!articleId) {
-    return;
-  }
-
-  const groupId = Number(firstString(prefsMap.groupId) ?? defaultGroupId) || defaultGroupId;
-  const ddmTemplateKey = firstString(prefsMap.ddmTemplateKey);
-  refs.set(articleId, {
-    articleId,
-    groupId,
-    ...(ddmTemplateKey ? {ddmTemplateKey} : {}),
-  });
-}
-
-async function fetchJournalArticleByUrlTitle(
-  gateway: LiferayGateway,
-  groupId: number,
-  urlTitle: string,
-): Promise<Record<string, unknown> | null> {
-  try {
-    const response = await safeGatewayGet<Record<string, unknown>>(
-      gateway,
-      `/api/jsonws/journal.journalarticle/get-article-by-url-title?groupId=${groupId}&urlTitle=${encodeURIComponent(urlTitle)}`,
-      'get-article-by-url-title',
-    );
-    if (!response.ok || hasJsonWsException(response.data)) {
-      return null;
-    }
-    return response.data ?? null;
-  } catch {
-    return null;
-  }
-}
-
-async function fetchLatestJournalArticle(
-  gateway: LiferayGateway,
-  groupId: number,
-  articleId: string,
-): Promise<Record<string, unknown> | null> {
-  try {
-    const response = await safeGatewayGet<Record<string, unknown>>(
-      gateway,
-      `/api/jsonws/journal.journalarticle/get-latest-article?groupId=${groupId}&articleId=${encodeURIComponent(articleId)}&status=0`,
-      'get-latest-article',
-    );
-    if (!response.ok || hasJsonWsException(response.data)) {
-      return null;
-    }
-    return response.data ?? null;
-  } catch {
-    return null;
-  }
-}
-
-async function fetchStructuredContentById(gateway: LiferayGateway, id: number): Promise<StructuredContent | null> {
-  if (id <= 0) {
-    return null;
-  }
-  const response = await safeGatewayGet<StructuredContent>(
-    gateway,
-    `/o/headless-delivery/v1.0/structured-contents/${id}`,
-    'fetch-structured-content-by-id',
-  );
-  return response.ok ? (response.data ?? null) : null;
-}
-
-async function fetchStructuredContentByUuid(
-  gateway: LiferayGateway,
-  groupId: number,
-  uuid: string,
-): Promise<StructuredContent | null> {
-  if (!uuid) {
-    return null;
-  }
-  const response = await safeGatewayGet<StructuredContent>(
-    gateway,
-    `/o/headless-delivery/v1.0/sites/${groupId}/structured-contents/by-uuid/${encodeURIComponent(uuid)}`,
-    'fetch-structured-content-by-uuid',
-  );
-  return response.ok ? (response.data ?? null) : null;
-}
-
-async function fetchContentStructureById(gateway: LiferayGateway, id: number): Promise<Record<string, unknown> | null> {
-  if (id <= 0) {
-    return null;
-  }
-  const response = await safeGatewayGet<Record<string, unknown>>(
-    gateway,
-    `/o/headless-delivery/v1.0/content-structures/${id}`,
-    'fetch-content-structure-by-id',
-  );
-  return response.ok ? (response.data ?? null) : null;
-}
-
-async function safeGatewayGet<T>(
-  gateway: LiferayGateway,
-  requestPath: string,
-  label: string,
-  requestOptions?: {headers?: Record<string, string>},
-) {
-  try {
-    const data = await gateway.getJson<T>(requestPath, label, requestOptions);
-    return {ok: true, status: 200, headers: new Headers(), body: '', data};
-  } catch {
-    return {ok: false, status: -1, headers: new Headers(), body: '', data: null};
-  }
-}
-
-async function resolveStructureSiteByKey(
-  gateway: LiferayGateway,
-  config: AppConfig,
-  apiClient: LiferayApiClient,
-  startSite: string,
-  structureKey: string,
-): Promise<{siteFriendlyUrl: string} | null> {
-  const siteChain = await buildResourceSiteChain(config, startSite, {apiClient, gateway});
-  for (const site of siteChain) {
-    const response = await safeGatewayGet<Record<string, unknown>>(
-      gateway,
-      `/o/data-engine/v2.0/sites/${site.siteId}/data-definitions/by-content-type/journal/by-data-definition-key/${encodeURIComponent(structureKey)}`,
-      'resolve-structure-site-by-key',
-    );
-    if (response.ok) {
-      return {siteFriendlyUrl: site.siteFriendlyUrl};
-    }
-  }
-  return null;
-}
-
-async function safeFetchGroupInfo(
-  config: AppConfig,
-  groupId: number,
-  dependencies: {apiClient: LiferayApiClient; gateway: LiferayGateway},
-) {
-  try {
-    return await fetchGroupInfo(config, groupId, dependencies);
-  } catch {
-    return null;
-  }
-}
-
-async function resolveTemplateSiteByKey(
-  config: AppConfig,
-  startSite: string,
-  templateKey: string,
-  dependencies: {apiClient: LiferayApiClient; gateway: LiferayGateway},
-): Promise<string | null> {
-  const siteChain = await buildResourceSiteChain(config, startSite, dependencies);
-  for (const candidate of siteChain) {
-    const site = await resolveResourceSite(config, candidate.siteFriendlyUrl, dependencies);
-    const templates = await listDdmTemplates(config, site, dependencies, {
-      includeCompanyFallback: candidate.siteFriendlyUrl === '/global',
-    });
-    if (templates.some((item) => matchesDdmTemplate(item, templateKey))) {
-      return candidate.siteFriendlyUrl;
-    }
-  }
-  return null;
-}
-
-function buildStructureExportPath(config: AppConfig, siteFriendlyUrl: string, key: string): string | undefined {
-  try {
-    return path.join(resolveArtifactSiteDir(config, 'structure', resolveSiteToken(siteFriendlyUrl)), `${key}.json`);
-  } catch {
-    return undefined;
-  }
-}
-
-function buildTemplateExportPath(config: AppConfig, siteFriendlyUrl: string, key: string): string | undefined {
-  try {
-    return path.join(resolveArtifactSiteDir(config, 'template', resolveSiteToken(siteFriendlyUrl)), `${key}.ftl`);
-  } catch {
-    return undefined;
-  }
-}
-
-async function enrichFragmentEntryExportPaths(
-  config: AppConfig,
-  gateway: LiferayGateway,
-  startSite: string,
-  entries: PageFragmentEntry[],
-  apiClient: LiferayApiClient,
-): Promise<void> {
-  const fragmentEntries = entries.filter((entry) => entry.type === 'fragment' && entry.fragmentKey);
-  if (fragmentEntries.length === 0) {
-    return;
-  }
-
-  const cache = new Map<string, {siteFriendlyUrl: string; exportPath: string} | null>();
-  for (const entry of fragmentEntries) {
-    const fragmentKey = entry.fragmentKey!;
-    if (!cache.has(fragmentKey)) {
-      cache.set(
-        fragmentKey,
-        await findFragmentExportPath(config, startSite, fragmentKey, {
-          apiClient,
-          gateway,
-        }),
-      );
-    }
-    const match = cache.get(fragmentKey);
-    if (match) {
-      entry.fragmentSiteFriendlyUrl = match.siteFriendlyUrl;
-      entry.fragmentExportPath = match.exportPath;
-    }
-  }
-}
-
-async function findFragmentExportPath(
-  config: AppConfig,
-  startSite: string,
-  fragmentKey: string,
-  dependencies: {apiClient: LiferayApiClient; gateway: LiferayGateway},
-): Promise<{siteFriendlyUrl: string; exportPath: string} | null> {
-  const baseDirs = await resolveFragmentSearchBaseDirs(config);
-  const siteChain = await safeBuildFragmentSiteChain(config, startSite, dependencies);
-  for (const site of siteChain) {
-    for (const baseDir of baseDirs) {
-      const siteDir = path.join(baseDir, 'sites', resolveSiteToken(site.siteFriendlyUrl));
-      const exportPath = await findFragmentDir(siteDir, fragmentKey);
-      if (exportPath) {
-        return {siteFriendlyUrl: site.siteFriendlyUrl, exportPath};
-      }
-    }
-  }
-  return null;
-}
-
-async function resolveFragmentSearchBaseDirs(config: AppConfig): Promise<string[]> {
-  const configured = resolveFragmentsBaseDir(config);
-  const candidates = [configured];
-  if (config.liferayDir && (await fs.pathExists(config.liferayDir))) {
-    const entries = await fs.readdir(config.liferayDir, {withFileTypes: true});
-    for (const entry of entries) {
-      if (entry.isDirectory() && /fragments$/i.test(entry.name)) {
-        candidates.push(path.join(config.liferayDir, entry.name));
-      }
-    }
-  }
-  return [...new Set(candidates)];
-}
-
-async function safeBuildFragmentSiteChain(
-  config: AppConfig,
-  startSite: string,
-  dependencies: {apiClient: LiferayApiClient; gateway: LiferayGateway},
-): Promise<Array<{siteFriendlyUrl: string}>> {
-  try {
-    return await buildResourceSiteChain(config, startSite, dependencies);
-  } catch {
-    return startSite === '/global'
-      ? [{siteFriendlyUrl: '/global'}]
-      : [{siteFriendlyUrl: startSite}, {siteFriendlyUrl: '/global'}];
-  }
-}
-
-async function findFragmentDir(root: string, fragmentKey: string): Promise<string | null> {
-  if (!(await fs.pathExists(root))) {
-    return null;
-  }
-  const queue = [root];
-  while (queue.length > 0) {
-    const current = queue.shift();
-    if (!current) {
-      continue;
-    }
-    const entries = await fs.readdir(current, {withFileTypes: true});
-    for (const entry of entries) {
-      const entryPath = path.join(current, entry.name);
-      if (!entry.isDirectory()) {
-        continue;
-      }
-      if (entry.name === fragmentKey && path.basename(path.dirname(entryPath)) === 'fragments') {
-        return entryPath;
-      }
-      queue.push(entryPath);
-    }
-  }
-  return null;
-}
-
-function inferContentStructureKey(value: Record<string, unknown> | null | undefined): string {
-  const explicit = String(value?.dataDefinitionKey ?? value?.key ?? '').trim();
-  if (explicit) {
-    return explicit;
-  }
-  const name = String(value?.name ?? '').trim();
-  return /^[A-Z0-9_]+$/.test(name) ? name : '';
-}
-
-/**
- * Fetch component page data (page element, metadata, fragment links).
- * Single responsibility: isolated API calls for content page data.
- */
-async function fetchComponentPageData(
-  gateway: LiferayGateway,
-  siteId: number,
-  canonicalFriendlyUrl: string,
-  plid: number,
-): Promise<{
-  pageElement: Record<string, unknown> | null;
-  pageMetadata: Record<string, unknown> | null;
-  rawFragmentLinks: Array<Record<string, unknown>>;
-}> {
-  const pageElement = await fetchSitePageElement(gateway, siteId, canonicalFriendlyUrl);
-  const pageMetadata = await tryFetchSitePageMetadata(gateway, siteId, canonicalFriendlyUrl);
-  const rawFragmentLinks = await tryFetchFragmentEntryLinks(gateway, siteId, plid);
-
-  return {pageElement, pageMetadata, rawFragmentLinks};
-}
-
-type TemplateInfo = {
-  widgetTemplateCandidates: string[];
-  displayPageTemplateCandidates: string[];
-  widgetHeadlessDefaultTemplate: string | undefined;
-  displayPageDefaultTemplate: string | undefined;
-};
-
-function extractTemplatesFromRenderedContents(renderedContents: Array<Record<string, unknown>>): TemplateInfo {
-  const widgetTemplateCandidates: string[] = [];
-  const displayPageTemplateCandidates: string[] = [];
-  let widgetHeadlessDefaultTemplate: string | undefined;
-  let displayPageDefaultTemplate: string | undefined;
-
-  for (const entry of renderedContents) {
-    const template = firstString(entry.contentTemplateName) ?? firstString(entry.contentTemplateId);
-    if (!template) {
-      continue;
-    }
-    const renderedUrl = firstString(entry.renderedContentURL) ?? '';
-    const markedAsDefault = Boolean(entry.markedAsDefault);
-    const isDisplayPage = renderedUrl.includes('/rendered-content-by-display-page/');
-    if (isDisplayPage) {
-      displayPageTemplateCandidates.push(template);
-      if (markedAsDefault && !displayPageDefaultTemplate) {
-        displayPageDefaultTemplate = template;
-      }
-    } else {
-      widgetTemplateCandidates.push(template);
-      if (markedAsDefault && !widgetHeadlessDefaultTemplate) {
-        widgetHeadlessDefaultTemplate = template;
-      }
-    }
-  }
-
-  return {
-    widgetTemplateCandidates,
-    displayPageTemplateCandidates,
-    widgetHeadlessDefaultTemplate,
-    displayPageDefaultTemplate,
-  };
-}
-
-function enrichJournalArticleWithStructuredContent(
-  summary: JournalArticleSummary,
-  structuredContent: StructuredContent,
-  ddmTemplateKey?: string,
-): void {
-  const record = asRecord(structuredContent);
-  const renderedContents = Array.isArray(record.renderedContents)
-    ? record.renderedContents.map((item) => asRecord(item))
-    : [];
-  const taxonomyCategoryBriefs = Array.isArray(record.taxonomyCategoryBriefs)
-    ? record.taxonomyCategoryBriefs.map((item) => asRecord(item))
-    : [];
-
-  // Extract template information from rendered contents
-  const templates = extractTemplatesFromRenderedContents(renderedContents);
-  const widgetDefaultTemplate = ddmTemplateKey || templates.widgetHeadlessDefaultTemplate;
-
-  const taxonomyCategoryNames = taxonomyCategoryBriefs
-    .map((item) => firstString(item.taxonomyCategoryName))
-    .filter((value): value is string => Boolean(value));
-
-  const priority = Number(record.priority);
-  const relatedContentsCount = Array.isArray(record.relatedContents) ? record.relatedContents.length : undefined;
-  const availableLanguages = Array.isArray(record.availableLanguages)
-    ? record.availableLanguages.map((item) => String(item)).filter(Boolean)
-    : [];
-
-  // Assign template properties
-  assignOptionalString(summary, 'widgetDefaultTemplate', widgetDefaultTemplate);
-  assignOptionalString(summary, 'widgetHeadlessDefaultTemplate', templates.widgetHeadlessDefaultTemplate);
-  assignOptionalString(summary, 'displayPageDefaultTemplate', templates.displayPageDefaultTemplate);
-  if (templates.widgetTemplateCandidates.length > 0) {
-    summary.widgetTemplateCandidates = templates.widgetTemplateCandidates;
-  }
-  if (templates.displayPageTemplateCandidates.length > 0) {
-    summary.displayPageTemplateCandidates = templates.displayPageTemplateCandidates;
-  }
-  if (renderedContents.length > 0) {
-    summary.renderedContents = renderedContents;
-  }
-  if (taxonomyCategoryBriefs.length > 0) {
-    summary.taxonomyCategoryBriefs = taxonomyCategoryBriefs;
-  }
-  if (taxonomyCategoryNames.length > 0) {
-    summary.taxonomyCategoryNames = taxonomyCategoryNames;
-  }
-  if (availableLanguages.length > 0) {
-    summary.availableLanguages = availableLanguages;
-  }
-
-  // Assign date properties using helper
-  assignOptionalString(summary, 'dateCreated', firstString(record.dateCreated));
-  assignOptionalString(summary, 'dateModified', firstString(record.dateModified));
-  assignOptionalString(summary, 'datePublished', firstString(record.datePublished));
-  assignOptionalString(
-    summary,
-    'expirationDate',
-    firstString(record.expirationDate) ?? firstString(record.dateExpired) ?? firstString(record.dateExpiration),
-  );
-  assignOptionalString(summary, 'reviewDate', firstString(record.reviewDate) ?? firstString(record.dateReview));
-  assignOptionalString(summary, 'description', firstString(record.description));
-  assignOptionalString(summary, 'externalReferenceCode', firstString(record.externalReferenceCode));
-  assignOptionalString(summary, 'uuid', firstString(record.uuid));
-
-  // Assign numeric properties
-  assignOptionalNumber(summary, 'siteId', Number(record.siteId));
-  assignOptionalNumber(summary, 'structuredContentFolderId', Number(record.structuredContentFolderId));
-  assignOptionalFiniteNumber(summary, 'priority', priority);
-
-  // Assign boolean properties
-  assignOptionalBoolean(summary, 'neverExpire', record.neverExpire);
-  assignOptionalBoolean(summary, 'subscribed', record.subscribed);
-  if (typeof relatedContentsCount === 'number') {
-    summary.relatedContentsCount = relatedContentsCount;
-  }
 }
